Create start.sh
This is utilised to create a new terminator session.
Initially creates a directory with the box name, as well as setting the IP address of the target. It has four panes. One full size on the left with three on the right.The bottom pane is in the binaries folder, with the other three in the box folder.